refactor: adds new folder organization
