+++ src/pages/Admin/AdminProducts/EditProduct/index.jsx
@@ -1,31 +1,34 @@
-import React, { useContext, useState } from "react";
+import React, { useContext, useEffect, useRef, useState } from "react";
 import "./styles.css";
 import { AddAPhotoSharp } from "@mui/icons-material";
-import { AppContext } from "./../../../contexts/AppContext";
+import { AppContext } from "../../../../contexts/AppContext";
 import axios from "axios";
 import { useDispatch, useSelector } from "react-redux";
-import { addItemProduct } from "../../../reducer/shoopingReducer";
+import { actualizarItemProduct } from "../../../../reducer/shoopingReducer";
 
-export default function AddProduct() {
+export default function EditProduct({ dataToEdit }) {
   const [imagePreview, setImagePreview] = useState(null);
   const state = useSelector((state) => state);
   const { user } = state.user;
   const dispatch = useDispatch();
-  const { setShowProductModal, setIsAddProduct, setLoader, onRequestShowAlert } =
+  const { setShowProductModal, setIsAddProduct, setLoader } =
     useContext(AppContext);
+  const productForm = useRef(null);
+  useEffect(() => {
+    setImagePreview(dataToEdit.url);
+    productForm.current.name.value = dataToEdit.name;
+    productForm.current.portion.value = dataToEdit.portion;
+    productForm.current.price.value = dataToEdit.price;
+    productForm.current.type.value = dataToEdit.type;
+  }, []);
 
   const handleSubmit = async (e) => {
     e.preventDefault();
-    e.target.file.parentElement.firstElementChild.style.border = "none";
+    // e.target.file.parentElement.firstElementChild.style.border = "none";
     e.target.name.style.borderColor = "#000000";
     e.target.portion.style.borderColor = "#000000";
     e.target.price.style.borderColor = "#000000";
     e.target.type.style.borderColor = "#000000";
-    if (e.target.file.files.length === 0) {
-      e.target.file.parentElement.firstElementChild.style.border =
-        "1px solid red";
-      return;
-    }
     if (e.target.name.value === "") {
       e.target.name.style.borderColor = "red";
       return;
@@ -44,33 +47,53 @@ export default function AddProduct() {
       return;
     }
     //here we do the form data and send to the backend
-    const product = new FormData();
-    product.append("image", e.target.file.files[0]);
-    product.append("name", e.target.name.value);
-    product.append("portion", e.target.portion.value);
-    product.append("price", e.target.price.value);
-    product.append("type", e.target.type.value);
     try {
-      const options = {
-        method: "POST",
-        url: "https://vegetanizando-api.vercel.app/product",
-        headers: {
-          "Content-Type":
-            "multipart/form-data; boundary=---011000010111000001101001",
-          "auth-token": user.token,
-        },
-        data: product,
-      };
+      let options;
+      if (e.target.file.files.length === 0) {
+        options = {
+          method: "PATCH",
+          url: `https://vegetanizando-api.vercel.app/product/${dataToEdit.id}`,
+          headers: {
+            "Content-Type": "application/json",
+            "auth-token": user.token,
+          },
+          data: {
+            name: e.target.name.value,
+            portion: e.target.portion.value,
+            price: e.target.price.value,
+            type: e.target.type.value,
+          },
+        };
+      } else {
+        const product = new FormData();
+        product.append("image", e.target.file.files[0]);
+        product.append("name", e.target.name.value);
+        product.append("portion", e.target.portion.value);
+        product.append("price", e.target.price.value);
+        product.append("type", e.target.type.value);
+        options = {
+          method: "PUT",
+          url: `https://vegetanizando-api.vercel.app/product/${dataToEdit.id}`,
+          headers: {
+            "Content-Type":
+              "multipart/form-data; boundary=---011000010111000001101001",
+            "auth-token": user.token,
+          },
+          data: product,
+        };
+      }
       setShowProductModal(false);
       setIsAddProduct(false);
       setLoader(true);
       let newProduct = await axios.request(options);
-      dispatch(addItemProduct(newProduct.data));
-
+      dispatch(
+        actualizarItemProduct({ id: dataToEdit.id, ...newProduct.data })
+      );
+      setLoader(true);
       setLoader(false);
     } catch (error) {
       setLoader(false);
-      onRequestShowAlert({ variant: 'warning', message: 'Ocorreu um erro', });
+      alert("ocorreu um error");
       console.log(error);
     }
   };
@@ -88,7 +111,11 @@ export default function AddProduct() {
   };
 
   return (
-    <form onSubmit={(e) => handleSubmit(e)} className="form-add-container">
+    <form
+      onSubmit={(e) => handleSubmit(e)}
+      className="form-add-container"
+      ref={productForm}
+    >
       <div className="form-align-content">
         <img className="form-img-item" src={imagePreview} alt="" />
         <input
@@ -140,7 +167,7 @@ export default function AddProduct() {
           type="button"
           value="Cancelar"
         />
-        <input className="form-btn" type="submit" value="Cadastrar" />
+        <input className="form-btn" type="submit" value="Atualizar" />
       </div>
     </form>
   );
